bot/pyproject.toml
[project]
name = "lms-bot"
version = "0.1.0"
description = "Telegram bot for LMS"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]